rename codon optimizer directories
