+++ src/codon_opt/optimizers/random_optimizer.py
@@ -1,4 +1,4 @@
-from src.qodon.optimizer import CodonOptimizer
+from src.codon_opt.codon_optimizer import CodonOptimizer
 
 
 class RandomOptimizer(CodonOptimizer):
