+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import os
 import pytest
-from src.qodon.optimizer import CodonOptimizer
+from src.codon_opt.codon_optimizer import CodonOptimizer
 from src.config.design_config import DesignConfig
 from src.config.fold_config import FoldConfig
 from src.rna_folding.rna_folder import RNAFolder
