+++ tests/test_cases/test_optimizers.py
@@ -7,7 +7,7 @@ def test_TFDE(caplog):
     Test Tensorflow Differential Evolution Optimizer runs correctly
 
     """
-    from src.qodon.optimizers.tf_differential_evo import TfDiffEv
+    from src.codon_opt.optimizers.tf_differential_evo import TfDiffEv
 
     testargs = [
         "-i",
@@ -37,7 +37,7 @@ def test_RAND(caplog):
     Test Random Optimizer runs correctly
 
     """
-    from src.qodon.optimizers.random_optimizer import RandomOptimizer
+    from src.codon_opt.optimizers.random_optimizer import RandomOptimizer
 
     testargs = [
         "-i",
@@ -67,7 +67,7 @@ def test_GA(caplog):
     Test Genetic Algorithm Optimizer runs correctly
 
     """
-    from src.qodon.optimizers.classical_ga import GeneticAlgorithm
+    from src.codon_opt.optimizers.classical_ga import GeneticAlgorithm
 
     testargs = [
         "-i",
@@ -97,7 +97,7 @@ def test_Metro(caplog):
     Test Metropolis Algorithm Optimizer runs correctly
 
     """
-    from src.qodon.optimizers.metro_optimizer import MetropolisOptimizer
+    from src.codon_opt.optimizers.metro_optimizer import MetropolisOptimizer
 
     testargs = [
         "-i",
@@ -127,7 +127,7 @@ def test_REMC(caplog):
     Test Replica Exchange Monte Carlo Optimizer runs correctly
 
     """
-    from src.qodon.optimizers.replica_exchange_mc import REMCOptimizer
+    from src.codon_opt.optimizers.replica_exchange_mc import REMCOptimizer
 
     testargs = [
         "-i",
@@ -157,7 +157,7 @@ def test_GA_MCSolver(caplog):
     Test Genetic Algorithm Optimizer runs correctly with Monte Carlo RNA folding
 
     """
-    from src.qodon.optimizers.classical_ga import GeneticAlgorithm
+    from src.codon_opt.optimizers.classical_ga import GeneticAlgorithm
 
     testargs = [
         "-i",
@@ -190,7 +190,7 @@ def test_GA_ExactSolver(caplog):
     folding
 
     """
-    from src.qodon.optimizers.classical_ga import GeneticAlgorithm
+    from src.codon_opt.optimizers.classical_ga import GeneticAlgorithm
 
     testargs = [
         "-i",
@@ -222,7 +222,7 @@ def test_convergence(caplog):
     Test that convergence will be achieved
 
     """
-    from src.qodon.optimizers.random_optimizer import RandomOptimizer
+    from src.codon_opt.optimizers.random_optimizer import RandomOptimizer
 
     testargs = [
         "-i",
